实现功能：推荐排行榜，分类排行榜

+++ server/src/main/java/zheye/server/entity/BookComment.java
@@ -25,9 +25,9 @@ public class BookComment extends Model<BookComment> {
 
     private String content;
 
-    private Integer bookId;
+    private Long bookId;
 
-    private Integer userId;
+    private Long userId;
 
     private String publishTime;
